create studio site start & move proxy server logic
