--- cli/commands/pm2/start.ts
@@ -1,32 +1,0 @@
-import { __ } from '@wordpress/i18n';
-import { PreviewCommandLoggerAction as LoggerAction } from 'common/logger-actions';
-import { startDaemon } from 'cli/lib/pm2-manager';
-import { Logger, LoggerError } from 'cli/logger';
-import { StudioArgv } from 'cli/types';
-
-export async function runCommand(): Promise< void > {
-	const logger = new Logger< LoggerAction >();
-
-	try {
-		logger.reportStart( LoggerAction.LOAD, __( 'Starting PM2 daemon…' ) );
-		await startDaemon();
-		logger.reportSuccess( __( 'PM2 daemon started' ) );
-	} catch ( error ) {
-		if ( error instanceof LoggerError ) {
-			logger.reportError( error );
-		} else {
-			const loggerError = new LoggerError( __( 'Failed to start PM2 daemon' ), error );
-			logger.reportError( loggerError );
-		}
-	}
-}
-
-export const registerCommand = ( yargs: StudioArgv ) => {
-	return yargs.command( {
-		command: 'start',
-		describe: __( 'Start the PM2 daemon' ),
-		handler: async () => {
-			await runCommand();
-		},
-	} );
-};
