--- cli/commands/pm2/stop.ts
@@ -1,33 +1,0 @@
-import { __ } from '@wordpress/i18n';
-import { PreviewCommandLoggerAction as LoggerAction } from 'common/logger-actions';
-import { stopDaemon } from 'cli/lib/pm2-manager';
-import { Logger, LoggerError } from 'cli/logger';
-import { StudioArgv } from 'cli/types';
-
-export async function runCommand(): Promise< void > {
-	const logger = new Logger< LoggerAction >();
-
-	try {
-		logger.reportStart( LoggerAction.LOAD, __( 'Stopping PM2 daemon…' ) );
-		await stopDaemon();
-		logger.reportSuccess( __( 'PM2 daemon stopped' ) );
-	} catch ( error ) {
-		if ( error instanceof LoggerError ) {
-			logger.reportError( error );
-		} else {
-			const loggerError = new LoggerError( __( 'Failed to stop PM2 daemon' ), error );
-			logger.reportError( loggerError );
-		}
-	}
-}
-
-export const registerCommand = ( yargs: StudioArgv ) => {
-	return yargs.command( {
-		command: 'stop',
-		describe: __( 'Stop the PM2 daemon' ),
-		handler: async () => {
-			await runCommand();
-		},
-	} );
-};
-
